+++ node/mdSchema.ts
@@ -35,6 +35,7 @@ export const COST_CENTER_FIELDS = [
   'organization',
   'phoneNumber',
   'businessDocument',
+  'customFields',
 ]
 export const COST_CENTER_SCHEMA_VERSION = 'v0.0.7'
 
@@ -132,6 +133,10 @@ export const schemas = [
           title: 'Created',
           format: 'date-time',
         },
+        customFields: {
+          type: ['array', 'null'],
+          title: 'Custom Fields',
+        },
       },
       'v-indexed': ['name', 'status', 'created'],
       'v-immediate-indexing': true,
@@ -166,6 +171,10 @@ export const schemas = [
         phoneNumber: {
           type: ['string', 'null'],
           title: 'Phone Number',
+        },
+        customFields: {
+          type: ['array', 'null'],
+          title: 'Custom Fields',
         },
       },
       'v-indexed': ['name', 'organization', 'businessDocument'],