+++ node/queries/getRole.ts
@@ -1,0 +1,12 @@
+import { print } from 'graphql'
+import gql from 'graphql-tag'
+
+export default print(gql`
+  query role($id: ID!) {
+    getRole(id: $id) {
+      id
+      name
+      slug
+    }
+  }
+`)